services/drought-management/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Drought management specific dependencies
numpy==1.24.3
pandas==2.1.4
scipy==1.11.4

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0

# Weather and climate data
requests==2.31.0

# Mathematical calculations for drought assessment
scikit-learn==1.3.2